fix(cli): clean up cmd with helpers
